4.19 to 4.23 finish

+++ part4/blogs/controllers/blogs.js
@@ -1,8 +1,5 @@
-const jwt = require('jsonwebtoken');
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
-const User = require('../models/user');
-const { errorHandler } = require('../utils/middleware');
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
@@ -47,17 +44,32 @@ blogsRouter.post('/', async (request, response,next) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id);
-    response.status(204).end();
+    
+    const blog = await Blog.findById(request.params.id);
+    if (blog.user.toString() !== request.user.id.toString()) {
+        return response.status(401).json({ error: 'unauthorized user' });
+    }else{
+        await Blog.findByIdAndDelete(request.params.id);
+        response.status(204).end();
+    }
+    
 });
 
 blogsRouter.put('/:id', async (request, response) => {
-    const blog = {
-        likes: request.body.likes
-    };
+    const user = request.user;
+    const blog = await Blog.findById(request.params.id);
+    if (blog.user.toString() !== user.id.toString()) {
+        return response.status(401).json({ error: 'unauthorized user' });
+    }else{
+        const blog = {
+            likes: request.body.likes
+        };
+    
+        const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true });
+        response.json(updatedBlog);
+    }
+    
 
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true });
-    response.json(updatedBlog);
 });
 
 module.exports = blogsRouter;
